backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.7.0,<3.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.7.0
torch>=2.0.0
transformers>=4.36.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.0
numpy>=1.24.0
